fix: rename `useMutations`
